Move migrate actions into core so they can be called by other plugins
