+++ x-pack/plugin/core/src/main/java/module-info.java
@@ -78,6 +78,8 @@ module org.elasticsearch.xcore {
     exports org.elasticsearch.xpack.core.inference.results;
     exports org.elasticsearch.xpack.core.inference;
     exports org.elasticsearch.xpack.core.logstash;
+    exports org.elasticsearch.xpack.core.migrate.action;
+    exports org.elasticsearch.xpack.core.migrate.task;
     exports org.elasticsearch.xpack.core.ml.action;
     exports org.elasticsearch.xpack.core.ml.annotations;
     exports org.elasticsearch.xpack.core.ml.autoscaling;
