+++ x-pack/plugin/core/src/main/java/org/elasticsearch/xpack/core/migrate/task/ReindexDataStreamEnrichedStatus.java
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-package org.elasticsearch.xpack.migrate.task;
+package org.elasticsearch.xpack.core.migrate.task;
 
 import org.elasticsearch.common.io.stream.StreamInput;
 import org.elasticsearch.common.io.stream.StreamOutput;
