+++ x-pack/plugin/core/src/test/java/org/elasticsearch/xpack/core/migrate/action/ReindexDatastreamIndexResponseTests.java
@@ -5,11 +5,11 @@
  * 2.0.
  */
 
-package org.elasticsearch.xpack.migrate.action;
+package org.elasticsearch.xpack.core.migrate.action;
 
 import org.elasticsearch.common.io.stream.Writeable;
 import org.elasticsearch.test.AbstractWireSerializingTestCase;
-import org.elasticsearch.xpack.migrate.action.ReindexDataStreamIndexAction.Response;
+import org.elasticsearch.xpack.core.migrate.action.ReindexDataStreamIndexAction.Response;
 
 public class ReindexDatastreamIndexResponseTests extends AbstractWireSerializingTestCase<Response> {
     @Override
