+++ x-pack/plugin/core/src/test/java/org/elasticsearch/xpack/core/migrate/task/ReindexDataStreamEnrichedStatusTests.java
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-package org.elasticsearch.xpack.migrate.task;
+package org.elasticsearch.xpack.core.migrate.task;
 
 import org.elasticsearch.ElasticsearchException;
 import org.elasticsearch.common.bytes.BytesReference;
